requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pinecone-client==2.2.4
pypdf2==3.0.1
langchain==0.0.350
langchain-google-genai==0.0.6
langchain-pinecone==0.0.3
tiktoken==0.5.2
sentence-transformers==2.2.2
numpy==1.24.4